+++ src/ft_printf/ft_formatter.c
@@ -6,18 +6,18 @@
 /*   By: lvan-gef <lvan-gef@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2023/02/12 17:09:47 by lvan-gef      #+#    #+#                 */
-/*   Updated: 2024/05/05 02:50:32 by lvan-gef      ########   odam.nl         */
+/*   Updated: 2024/11/25 22:19:27 by lvan-gef      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../include/ft_printf.h"
 
-int	to_hex(unsigned long int size, size_t upper, size_t base)
+ssize_t	to_hex(unsigned long int size, size_t upper, size_t base)
 {
-	size_t	remainder;
+	ssize_t	remainder;
 	char	letter;
-	size_t	counter;
-	int		result;
+	ssize_t	counter;
+	ssize_t	result;
 
 	letter = 'a';
 	counter = 0;
@@ -29,21 +29,21 @@ int	to_hex(unsigned long int size, size_t upper, size_t base)
 	if (result < 0)
 		return (result);
 	counter += result;
-	remainder = size % base;
+	remainder = (ssize_t)(size % base);
 	if (remainder < 10)
-		result = write_ascii(remainder + '0');
+		result = write_ascii((int)remainder + '0');
 	else
-		result = write_ascii(remainder - 10 + letter);
+		result = write_ascii((int)remainder - 10 + letter);
 	if (result < 0)
 		return (result);
 	counter += result;
 	return (counter);
 }
 
-int	format_int(int size, size_t upper, size_t base)
+ssize_t	format_int(int size, size_t upper, size_t base)
 {
 	int		result;
-	int		counter;
+	ssize_t	counter;
 	size_t	new_size;
 
 	result = 0;
@@ -63,13 +63,13 @@ int	format_int(int size, size_t upper, size_t base)
 	return (result);
 }
 
-int	format_ptr(void *ptr, size_t upper, size_t base)
+ssize_t	format_ptr(void *ptr, size_t upper, size_t base)
 {
-	long int	addr;
-	int			counter;
-	int			result;
+	unsigned long int	addr;
+	ssize_t				counter;
+	int					result;
 
-	addr = (long int)ptr;
+	addr = (unsigned long int)ptr;
 	result = write_str("0x");
 	if (result < 0)
 		return (result);
