+++ src/str/ft_substr.c
@@ -6,7 +6,7 @@
 /*   By: lvan-gef <lvan-gef@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2022/10/25 17:02:55 by lvan-gef      #+#    #+#                 */
-/*   Updated: 2024/05/04 23:17:24 by lvan-gef      ########   odam.nl         */
+/*   Updated: 2024/11/25 22:09:01 by lvan-gef      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,14 +14,14 @@
 
 char	*ft_substr(char const *s, unsigned int start, size_t len)
 {
-	char			*str;
-	unsigned int	str_len;
+	char	*str;
+	size_t	str_len;
 
 	if (!s)
 		return (NULL);
 	str_len = ft_strlen(s);
 	if (start > str_len)
-		start = str_len;
+		start = (unsigned int)str_len;
 	if (len > str_len)
 		len = str_len;
 	if (len > str_len - start)
